use dedicated factory for input dto creation
