+++ src/DTO/InputDto.js
@@ -62,44 +62,6 @@ export default class InputDto {
   }
 
   /**
-   * Create input dto from GitHub Action input reader.
-   *
-   * @param {(name: string) => string} getInput
-   * @returns {InputDto}
-   */
-  static fromActionInput(getInput) {
-    /**
-     * @param {string} name
-     * @param {string} fallback
-     * @returns {string}
-     */
-    const read = (name, fallback) => {
-      const value = getInput(name)
-      return value === "" ? fallback : value
-    }
-
-    return new InputDto({
-      url: read("url", "http://localhost"),
-      databaseType: read("database_type", "pdo.mysql"),
-      databaseHost: read("database_host", "127.0.0.1"),
-      databaseUser: read("database_user", ""),
-      databasePassword: read("database_password", ""),
-      databaseName: read("database_name", "icms"),
-      databaseCharset: read("database_charset", "utf8"),
-      databaseCollation: read("database_collation", "utf8_general_ci"),
-      databasePrefix: read("database_prefix", "icms"),
-      databasePort: read("database_port", "3306"),
-      adminName: read("admin_name", "icms"),
-      adminLogin: read("admin_login", "icms"),
-      adminPass: read("admin_pass", "icms"),
-      adminEmail: read("admin_email", "noreply@impresscms.dev"),
-      language: read("language", "english"),
-      appKey: read("app_key", ""),
-      path: read("path", ".")
-    })
-  }
-
-  /**
    * @returns {string}
    */
   get url() {
